Work of the day

+++ src/syscall/syscall64_64/syscall64_64.h
@@ -7,7 +7,7 @@ extern "C" {
 
 #include "sysnum.h"
 
-extern int syscall64_64(Sysnum no, uint64_t p0, uint64_t p1, uint64_t p2, uint64_t p3, uint64_t p4, uint64_t p5);
+extern long syscall64_64(Sysnum no, uint64_t p0, uint64_t p1, uint64_t p2, uint64_t p3, uint64_t p4, uint64_t p5);
 
 #endif
